+++ project/GenSafeStyles.scala
@@ -20,6 +20,8 @@ import scala.io.Source
 
 object GenSafeStyles {
 
+  val generatorSource = new File("GenSafeStyles.scala")
+
   def translateLine(traitName: String)(line: String): String =
     line.replaceAllLiterally("with TestRegistration", "with SafeTestRegistration")
         .replaceAllLiterally("Any /* Assertion */", "Assertion")
@@ -35,43 +37,45 @@ object GenSafeStyles {
 
   def translateFile(targetDir: File, fileName: String, sourceFileName: String, scalaVersion: String, scalaJS: Boolean, translateFun: String => String): File = {
     val outputFile = new File(targetDir, fileName)
-    val outputWriter = new BufferedWriter(new FileWriter(outputFile))
-    try {
-      val lines = Source.fromFile(new File(sourceFileName)).getLines.toList
-      var skipMode = false
-      for (line <- lines) {
-        val mustLine: String =
-          if (scalaJS) {
-            if (line.trim == "// SKIP-SCALATESTJS-START") {
-              skipMode = true
-              ""
-            }
-            else if (line.trim == "// SKIP-SCALATESTJS-END") {
-              skipMode = false
-              ""
-            }
-            else if (!skipMode) {
-              if (line.trim.startsWith("//SCALATESTJS-ONLY "))
-                translateFun(line.substring(line.indexOf("//SCALATESTJS-ONLY ") + 19))
+    if (!outputFile.exists || generatorSource.lastModified > outputFile.lastModified) {
+      val outputWriter = new BufferedWriter(new FileWriter(outputFile))
+      try {
+        val lines = Source.fromFile(new File(sourceFileName)).getLines.toList
+        var skipMode = false
+        for (line <- lines) {
+          val mustLine: String =
+            if (scalaJS) {
+              if (line.trim == "// SKIP-SCALATESTJS-START") {
+                skipMode = true
+                ""
+              }
+              else if (line.trim == "// SKIP-SCALATESTJS-END") {
+                skipMode = false
+                ""
+              }
+              else if (!skipMode) {
+                if (line.trim.startsWith("//SCALATESTJS-ONLY "))
+                  translateFun(line.substring(line.indexOf("//SCALATESTJS-ONLY ") + 19))
+                else
+                  translateFun(line)
+              }
               else
-                translateFun(line)
+                ""
             }
             else
-              ""
-          }
-          else
-            translateFun(line)
+              translateFun(line)
 
-        outputWriter.write(mustLine)
-        outputWriter.newLine()
+          outputWriter.write(mustLine)
+          outputWriter.newLine()
+        }
       }
-      outputFile
+      finally {
+        outputWriter.flush()
+        outputWriter.close()
+        println("Generated " + outputFile.getAbsolutePath)
+      }
     }
-    finally {
-      outputWriter.flush()
-      outputWriter.close()
-      println("Generated " + outputFile.getAbsolutePath)
-    }
+    outputFile
   }
 
   def genMainImpl(targetDir: File, version: String, scalaVersion: String, scalaJS: Boolean): Seq[File] = {
